More reliable mine explosion damage
Previously it was actually possible to trigger a mine yet evade the
damage, either because the triggering tank was already within the
expanding area and thus never triggered the `body_entered` signal, or
because the triggering tank simply outran the expanding area.

In any case, we now use a single area to define the explosion radius,
and get all bodies within it; we also get all area overlapping it as
well in order to find and detonate other nearby mines.